refactor(hickey): move isActiveNic out of the formatting module
Hickey F1 — metrics.ts is the formatting/derivation module (free of
Solid/DOM); isActiveNic is a NIC activity-classification predicate, a
distinct volatility axis. Lowy cross-validation preferred a thin
predicates module over inlining (keeps the pure unit test). Extracted to
client/nic.ts with its test in client/nic.test.ts.
