+++ packages/app/src/client/metrics.test.ts
@@ -1,11 +1,10 @@
 import { describe, expect, it } from "bun:test";
-import type { NetInterface, SystemInfo } from "../common/surface";
+import type { SystemInfo } from "../common/surface";
 import {
   averageCoreUsage,
   formatBytes,
   formatThroughput,
   formatUptime,
-  isActiveNic,
   memGb,
   memPct,
 } from "./metrics";
@@ -78,36 +77,6 @@ describe("formatThroughput", () => {
   it("appends a per-second suffix to the byte size", () => {
     expect(formatThroughput(0)).toBe("0 B/s");
     expect(formatThroughput(1_200_000)).toBe("1.2 MB/s");
-  });
-});
-
-describe("isActiveNic", () => {
-  const nic = (over: Partial<NetInterface> = {}): NetInterface => ({
-    rxBytes: 0,
-    txBytes: 0,
-    rxRate: 0,
-    txRate: 0,
-    ...over,
-  });
-
-  it("is inactive when both rates are zero", () => {
-    expect(isActiveNic(nic())).toBe(false);
-  });
-
-  it("is active on receive throughput", () => {
-    expect(isActiveNic(nic({ rxRate: 1 }))).toBe(true);
-  });
-
-  it("is active on transmit throughput", () => {
-    expect(isActiveNic(nic({ txRate: 1 }))).toBe(true);
-  });
-
-  it("is inactive for a missing NIC (mid-tick churn)", () => {
-    expect(isActiveNic(undefined)).toBe(false);
-  });
-
-  it("ignores cumulative byte totals — only live rates count", () => {
-    expect(isActiveNic(nic({ rxBytes: 1e9, txBytes: 1e9 }))).toBe(false);
   });
 });
