+++ packages/app/src/client/metrics.ts
@@ -9,7 +9,7 @@
  * against one module.
  */
 
-import type { NetInterface, SystemInfo } from "../common/surface";
+import type { SystemInfo } from "../common/surface";
 
 export { averageCoreUsage, memPct } from "../common/metrics";
 
@@ -62,16 +62,4 @@ export function formatBytes(bytes: number): string {
 /** Throughput rendered as a per-second byte rate, e.g. "1.2 MB/s". */
 export function formatThroughput(bytesPerSec: number): string {
   return `${formatBytes(bytesPerSec)}/s`;
-}
-
-/** A NIC counts as "active" only when it moved bytes in the last poll window
- *  — cumulative lifetime totals don't matter. Idle interfaces (both rates 0)
- *  are collapsed by default in the network strip so the handful of busy NICs
- *  aren't buried under the dozens of always-zero virtual ones (utunN, anpiN,
- *  …) a typical host carries. `undefined` (a NIC seen mid-tick churn) is
- *  treated as inactive. */
-export function isActiveNic(
-  nic: Pick<NetInterface, "rxRate" | "txRate"> | undefined,
-): boolean {
-  return (nic?.rxRate ?? 0) > 0 || (nic?.txRate ?? 0) > 0;
 }
